Remove DB-specific folders
