+++ packages/core/src/indexing/__tests__/sqlite-index-api.test.ts
@@ -1,27 +1,27 @@
 import { jest } from '@jest/globals'
-import { PostgresIndexApi } from '../postgres-index-api.js'
+import tmp from 'tmp-promise'
 import { StreamID } from '@ceramicnetwork/streamid'
 import knex, { Knex } from 'knex'
-import pgSetup from '@databases/pg-test/jest/globalSetup'
-import pgTeardown from '@databases/pg-test/jest/globalTeardown'
-import { asTableName } from '../../as-table-name.util.js'
-import { IndexQueryNotAvailableError } from '../../index-query-not-available.error.js'
+import { IndexQueryNotAvailableError } from '../index-query-not-available.error.js'
+import { asTableName } from '../as-table-name.util.js'
 import { Model } from '@ceramicnetwork/stream-model'
 import { LoggerProvider, Networks } from '@ceramicnetwork/common'
 import { CID } from 'multiformats/cid'
-import { INDEXED_MODEL_CONFIG_TABLE_NAME, IndexModelArgs } from '../../database-index-api.js'
 import {
-  COMMON_TABLE_STRUCTURE,
-  RELATION_COLUMN_STRUCTURE,
-  CONFIG_TABLE_MODEL_INDEX_STRUCTURE,
-} from '../migrations/cdb-schema-verification.js'
-import { readCsvFixture } from '../../__tests__/read-csv-fixture.util.js'
-import { CONFIG_TABLE_NAME } from '../../config.js'
+  INDEXED_MODEL_CONFIG_TABLE_NAME,
+  IndexModelArgs,
+  SqliteIndexApi,
+  asTimestamp,
+} from '../database-index-api.js'
+import { DatabaseType } from '../migrations/1-create-model-table.js'
+import { STRUCTURES } from '../migrations/cdb-schema-verification.js'
+import { readCsvFixture } from './read-csv-fixture.util.js'
 
+const STRUCTURE = STRUCTURES[DatabaseType.SQLITE]
 const STREAM_ID_A = 'kjzl6cwe1jw145m7jxh4jpa6iw1ps3jcjordpo81e0w04krcpz8knxvg5ygiabd'
 const STREAM_ID_B = 'kjzl6cwe1jw147dvq16zluojmraqvwdmbh61dx9e0c59i344lcrsgqfohexp60s'
 const CONTROLLER = 'did:key:foo'
-const STREAM_TEST_DATA_PROFILE_A = {
+const STREAM_TEST_DATA_PROFILE = {
   id: 'bea4d783-6496-4a28-bf02-6603e56edf0a',
   name: 'Joeline Bradshaw',
   address: 'Attitudes Road 5270, Adena, Dominica, 509754',
@@ -31,22 +31,11 @@ const STREAM_TEST_DATA_PROFILE_A = {
     dark_mode: true,
   },
 }
-const STREAM_TEST_DATA_PROFILE_B = {
-  id: '509a7076-b4db-476c-8800-88db91c34796',
-  name: 'Rainier Enciso',
-  address: 'Activation St 7366, Krasnoyarsk, Mauritius, 875222',
-  birthDate: '29.12.1974',
-  email: 'dia_demelodked@toddler.jq',
-  settings: {
-    dark_mode: false,
-  },
-}
-const FAKE_CID_A = CID.parse('bafybeig6xv5nwphfmvcnektpnojts33jqcuam7bmye2pb54adnrtccjlsu')
-const FAKE_CID_B = CID.parse('bafybeig6xv5nwphfmvcnektpnojts44jqcuam7bmye2pb54adnrtccjlsu')
+const FAKE_CID = CID.parse('bafybeig6xv5nwphfmvcnektpnojts33jqcuam7bmye2pb54adnrtccjlsu')
 const logger = new LoggerProvider().getDiagnosticsLogger()
 
+let tmpFolder: tmp.DirectoryResult
 let dbConnection: Knex
-jest.setTimeout(150000) // 2.5mins timeout for initial docker fetch+init
 
 function modelsToIndexArgs(models: Array<StreamID>): Array<IndexModelArgs> {
   return models.map((model) => {
@@ -54,39 +43,28 @@ function modelsToIndexArgs(models: Array<StreamID>): Array<IndexModelArgs> {
   })
 }
 
-beforeAll(async () => {
-  await pgSetup()
-})
-
 beforeEach(async () => {
+  tmpFolder = await tmp.dir({ unsafeCleanup: true })
+  const filename = `${tmpFolder.path}/tmp-ceramic.sqlite`
   dbConnection = knex({
-    client: 'pg',
-    connection: process.env.DATABASE_URL,
+    client: 'sqlite3',
+    useNullAsDefault: true,
+    connection: {
+      filename: filename,
+    },
   })
 })
 
 afterEach(async () => {
-  await dropTables()
   await dbConnection.destroy()
-})
-
-export async function dropTables() {
-  await dbConnection.schema.dropTableIfExists(CONFIG_TABLE_NAME)
-  await dbConnection.schema.dropTableIfExists(INDEXED_MODEL_CONFIG_TABLE_NAME)
-  await dbConnection.schema.dropTableIfExists(Model.MODEL.toString())
-  await dbConnection.schema.dropTableIfExists(STREAM_ID_A)
-  await dbConnection.schema.dropTableIfExists(STREAM_ID_B)
-}
-
-afterAll(async () => {
-  await pgTeardown()
+  await tmpFolder.cleanup()
 })
 
 describe('init', () => {
   describe('create tables', () => {
     test('create new table from scratch', async () => {
       const modelToIndex = StreamID.fromString(STREAM_ID_A)
-      const indexApi = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+      const indexApi = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
       await indexApi.init()
       await indexApi.indexModels(modelsToIndexArgs([modelToIndex]))
       const created = await indexApi.tablesManager.listMidTables()
@@ -100,12 +78,16 @@ describe('init', () => {
       ).resolves.not.toThrow()
 
       // Also manually check MID table structure
-      let columns = await dbConnection.table(asTableName(modelToIndex)).columnInfo()
-      expect(JSON.stringify(columns)).toEqual(JSON.stringify(COMMON_TABLE_STRUCTURE))
+      const midColumns = await dbConnection.table(asTableName(modelToIndex)).columnInfo()
+      expect(JSON.stringify(midColumns)).toEqual(JSON.stringify(STRUCTURE.COMMON_TABLE))
 
       // Also manually check config table structure
-      columns = await dbConnection.table(asTableName(INDEXED_MODEL_CONFIG_TABLE_NAME)).columnInfo()
-      expect(JSON.stringify(columns)).toEqual(JSON.stringify(CONFIG_TABLE_MODEL_INDEX_STRUCTURE))
+      const configTableColumns = await dbConnection
+        .table(asTableName(INDEXED_MODEL_CONFIG_TABLE_NAME))
+        .columnInfo()
+      expect(JSON.stringify(configTableColumns)).toEqual(
+        JSON.stringify(STRUCTURE.CONFIG_TABLE_MODEL_INDEX)
+      )
     })
 
     test('create new table with relations', async () => {
@@ -115,7 +97,7 @@ describe('init', () => {
           relations: { fooRelation: { type: 'account' } },
         },
       ]
-      const indexApi = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+      const indexApi = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
       await indexApi.init()
       await indexApi.indexModels(indexModelsArgs)
       const created = await indexApi.tablesManager.listMidTables()
@@ -126,55 +108,55 @@ describe('init', () => {
 
       // Also manually check table structure
       const columns = await dbConnection.table(asTableName(indexModelsArgs[0].model)).columnInfo()
-      const expectedTableStructure = Object.assign({}, COMMON_TABLE_STRUCTURE, {
-        fooRelation: RELATION_COLUMN_STRUCTURE,
+      const expectedTableStructure = Object.assign({}, STRUCTURE.COMMON_TABLE, {
+        fooRelation: STRUCTURE.RELATION_COLUMN,
       })
       expect(JSON.stringify(columns)).toEqual(JSON.stringify(expectedTableStructure))
     })
 
     test('table creation is idempotent', async () => {
-      const modelsToIndex = [Model.MODEL, StreamID.fromString(STREAM_ID_A)]
-      const indexApi = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+      const modelsToIndex = [StreamID.fromString(STREAM_ID_A), Model.MODEL]
+      const indexApi = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
       await indexApi.init()
       await indexApi.indexModels(modelsToIndexArgs(modelsToIndex))
       // Index the same models again to make sure we don't error trying to re-create the tables
       await indexApi.indexModels(modelsToIndexArgs(modelsToIndex))
       const created = await indexApi.tablesManager.listMidTables()
-      const tableNames = modelsToIndex.map(asTableName)
+      const tableNames = modelsToIndex.map((m) => `${m.toString()}`)
       expect(created.sort()).toEqual(tableNames.sort())
     })
 
     test('create new table with existing ones', async () => {
       // First init with one model
       const modelsA = [StreamID.fromString(STREAM_ID_A)]
-      const indexApiA = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+      const indexApiA = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
       await indexApiA.init()
       await indexApiA.indexModels(modelsToIndexArgs(modelsA))
       const createdA = await indexApiA.tablesManager.listMidTables()
-      const tableNamesA = modelsA.map(asTableName)
+      const tableNamesA = modelsA.map((m) => `${m.toString()}`)
       expect(createdA.sort()).toEqual(tableNamesA.sort())
 
       // Next add another one
       const modelsB = [...modelsA, StreamID.fromString(STREAM_ID_B)]
-      const indexApiB = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+      const indexApiB = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
       await indexApiB.indexModels(modelsToIndexArgs(modelsB))
       const createdB = await indexApiB.tablesManager.listMidTables()
-      const tableNamesB = modelsB.map(asTableName)
+      const tableNamesB = modelsB.map((m) => `${m.toString()}`)
       expect(createdB.sort()).toEqual(tableNamesB.sort())
     })
 
     test('checks or persists the network used for indexing', async () => {
-      const indexApiA = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+      const indexApiA = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
       await indexApiA.init()
 
       // Throws if initialized with a different network
-      const indexApiB = new PostgresIndexApi(dbConnection, true, logger, Networks.MAINNET)
+      const indexApiB = new SqliteIndexApi(dbConnection, true, logger, Networks.MAINNET)
       await expect(indexApiB.init()).rejects.toThrow(
         'Initialization failed for config table: ceramic_config. The database is configured to use the network inmemory but the current network is mainnet.'
       )
 
       // Does not throw if initialized with the stored network
-      const indexApiC = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+      const indexApiC = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
       await expect(indexApiC.init()).resolves.toBeUndefined()
     })
   })
@@ -192,26 +174,19 @@ describe('init', () => {
      */
     test('Can manually create table that passes validation', async () => {
       const modelToIndex = StreamID.fromString(STREAM_ID_A)
-      const tableName = asTableName(modelToIndex)
-      const indexApi = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+      const indexApi = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
       await indexApi.init()
 
       // Create the table in the database with all expected fields but one (leaving off 'updated_at')
-      await dbConnection.schema.createTable(tableName, (table) => {
-        // create unique index name <64 chars that are still capable of being referenced to MID table
-        const indexName = tableName.substring(tableName.length - 10)
-
-        table
-          .string('stream_id')
-          .primary(`idx_${indexName}_pkey`)
-          .unique(`constr_${indexName}_unique`)
+      await dbConnection.schema.createTable(asTableName(modelToIndex), (table) => {
+        table.string('stream_id', 1024).primary().unique().notNullable()
         table.string('controller_did', 1024).notNullable()
-        table.jsonb('stream_content').notNullable()
+        table.string('stream_content').notNullable()
         table.string('tip').notNullable()
-        table.dateTime('last_anchored_at').nullable()
-        table.dateTime('first_anchored_at').nullable()
-        table.dateTime('created_at').notNullable().defaultTo(dbConnection.fn.now())
-        table.dateTime('updated_at').notNullable().defaultTo(dbConnection.fn.now())
+        table.integer('last_anchored_at').nullable()
+        table.integer('first_anchored_at').nullable()
+        table.integer('created_at').notNullable()
+        table.integer('updated_at').notNullable()
       })
 
       await expect(
@@ -221,26 +196,18 @@ describe('init', () => {
 
     test('Fail table validation', async () => {
       const modelToIndex = StreamID.fromString(STREAM_ID_A)
-      const tableName = asTableName(modelToIndex)
-
-      const indexApi = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+      const indexApi = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
       await indexApi.init()
 
       // Create the table in the database with all expected fields but one (leaving off 'updated_at')
-      await dbConnection.schema.createTable(tableName, (table) => {
-        // create unique index name <64 chars that are still capable of being referenced to MID table
-        const indexName = tableName.substring(tableName.length - 10)
-
-        table
-          .string('stream_id')
-          .primary(`idx_${indexName}_pkey`)
-          .unique(`constr_${indexName}_unique`)
+      await dbConnection.schema.createTable(asTableName(modelToIndex), (table) => {
+        table.string('stream_id', 1024).primary().unique().notNullable()
         table.string('controller_did', 1024).notNullable()
-        table.jsonb('stream_content').notNullable()
+        table.string('stream_content').notNullable()
         table.string('tip').notNullable()
-        table.dateTime('last_anchored_at').nullable()
-        table.dateTime('first_anchored_at').nullable()
-        table.dateTime('created_at').notNullable().defaultTo(dbConnection.fn.now())
+        table.integer('last_anchored_at').nullable()
+        table.integer('first_anchored_at').nullable()
+        table.integer('created_at').notNullable()
       })
 
       await expect(
@@ -250,7 +217,6 @@ describe('init', () => {
 
     test('Fail table validation if relation column missing', async () => {
       const modelToIndex = StreamID.fromString(STREAM_ID_A)
-      const tableName = asTableName(modelToIndex)
       const indexModelsArgs: Array<IndexModelArgs> = [
         {
           model: StreamID.fromString(STREAM_ID_A),
@@ -258,25 +224,19 @@ describe('init', () => {
         },
       ]
 
-      const indexApi = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+      const indexApi = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
       await indexApi.init()
 
       // Create the table in the database with all expected fields but one (leaving off 'updated_at')
-      await dbConnection.schema.createTable(tableName, (table) => {
-        // create unique index name <64 chars that are still capable of being referenced to MID table
-        const indexName = tableName.substring(tableName.length - 10)
-
-        table
-          .string('stream_id')
-          .primary(`idx_${indexName}_pkey`)
-          .unique(`constr_${indexName}_unique`)
+      await dbConnection.schema.createTable(asTableName(modelToIndex), (table) => {
+        table.string('stream_id', 1024).primary().unique().notNullable()
         table.string('controller_did', 1024).notNullable()
-        table.jsonb('stream_content').notNullable()
+        table.string('stream_content').notNullable()
         table.string('tip').notNullable()
-        table.dateTime('last_anchored_at').nullable()
-        table.dateTime('first_anchored_at').nullable()
-        table.dateTime('created_at').notNullable().defaultTo(dbConnection.fn.now())
-        table.dateTime('updated_at').notNullable().defaultTo(dbConnection.fn.now())
+        table.integer('last_anchored_at').nullable()
+        table.integer('first_anchored_at').nullable()
+        table.integer('created_at').notNullable()
+        table.integer('updated_at').notNullable()
       })
 
       await expect(indexApi.tablesManager.verifyTables(indexModelsArgs)).rejects.toThrow(
@@ -291,7 +251,7 @@ describe('close', () => {
     const fauxDbConnection = {
       destroy: jest.fn(),
     } as unknown as Knex
-    const indexApi = new PostgresIndexApi(fauxDbConnection, true, logger, Networks.INMEMORY)
+    const indexApi = new SqliteIndexApi(fauxDbConnection, true, logger, Networks.INMEMORY)
     await indexApi.close()
     expect(fauxDbConnection.destroy).toBeCalled()
   })
@@ -301,15 +261,15 @@ describe('close', () => {
  * Difference between `a` and `b` timestamps is less than or equal to `deltaS`.
  */
 function closeDates(a: Date, b: Date, deltaS = 1) {
-  const aSeconds = a.getSeconds()
-  const bSeconds = b.getSeconds()
+  const aSeconds = asTimestamp(a)
+  const bSeconds = asTimestamp(b)
   return Math.abs(aSeconds - bSeconds) <= deltaS
 }
 
 describe('indexModels', () => {
   test('populates the INDEXED_MODEL_CONFIG_TABLE_NAME table on indexModels()', async () => {
     const modelsToIndex = [StreamID.fromString(STREAM_ID_A), Model.MODEL]
-    const indexApi = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+    const indexApi = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
     await indexApi.init()
     await indexApi.indexModels(modelsToIndexArgs(modelsToIndex))
 
@@ -320,18 +280,18 @@ describe('indexModels', () => {
     ).toEqual([
       {
         model: 'kjzl6cwe1jw145m7jxh4jpa6iw1ps3jcjordpo81e0w04krcpz8knxvg5ygiabd',
-        is_indexed: true,
+        is_indexed: 1,
       },
       {
         model: 'kh4q0ozorrgaq2mezktnrmdwleo1d',
-        is_indexed: true,
+        is_indexed: 1,
       },
     ])
   })
 
   test('updates the INDEXED_MODEL_CONFIG_TABLE_NAME table on stopIndexingModels()', async () => {
     const modelsToIndex = [StreamID.fromString(STREAM_ID_A), Model.MODEL]
-    const indexApi = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+    const indexApi = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
     await indexApi.init()
     await indexApi.indexModels(modelsToIndexArgs(modelsToIndex))
     await indexApi.stopIndexingModels([StreamID.fromString(STREAM_ID_A)])
@@ -343,18 +303,18 @@ describe('indexModels', () => {
     ).toEqual([
       {
         model: 'kjzl6cwe1jw145m7jxh4jpa6iw1ps3jcjordpo81e0w04krcpz8knxvg5ygiabd',
-        is_indexed: false,
+        is_indexed: 0,
       },
       {
         model: 'kh4q0ozorrgaq2mezktnrmdwleo1d',
-        is_indexed: true,
+        is_indexed: 1,
       },
     ])
   })
 
   test('re-indexing models', async () => {
     const modelsToIndex = [StreamID.fromString(STREAM_ID_A), Model.MODEL]
-    const indexApi = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+    const indexApi = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
     await indexApi.init()
 
     await indexApi.indexModels(modelsToIndexArgs(modelsToIndex))
@@ -365,11 +325,11 @@ describe('indexModels', () => {
     ).toEqual([
       {
         model: 'kjzl6cwe1jw145m7jxh4jpa6iw1ps3jcjordpo81e0w04krcpz8knxvg5ygiabd',
-        is_indexed: true,
+        is_indexed: 1,
       },
       {
         model: 'kh4q0ozorrgaq2mezktnrmdwleo1d',
-        is_indexed: true,
+        is_indexed: 1,
       },
     ])
 
@@ -381,11 +341,11 @@ describe('indexModels', () => {
     ).toEqual([
       {
         model: 'kjzl6cwe1jw145m7jxh4jpa6iw1ps3jcjordpo81e0w04krcpz8knxvg5ygiabd',
-        is_indexed: false,
+        is_indexed: 0,
       },
       {
         model: 'kh4q0ozorrgaq2mezktnrmdwleo1d',
-        is_indexed: true,
+        is_indexed: 1,
       },
     ])
 
@@ -397,22 +357,22 @@ describe('indexModels', () => {
     ).toEqual([
       {
         model: 'kjzl6cwe1jw145m7jxh4jpa6iw1ps3jcjordpo81e0w04krcpz8knxvg5ygiabd',
-        is_indexed: true,
+        is_indexed: 1,
       },
       {
         model: 'kh4q0ozorrgaq2mezktnrmdwleo1d',
-        is_indexed: true,
+        is_indexed: 1,
       },
     ])
   })
 
   test('modelsToIndex is properly populated after init()', async () => {
     const modelsToIndex = [StreamID.fromString(STREAM_ID_A), Model.MODEL]
-    const indexApi = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+    const indexApi = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
     await indexApi.init()
     await indexApi.indexModels(modelsToIndexArgs(modelsToIndex))
 
-    const anotherIndexApi = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+    const anotherIndexApi = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
     await anotherIndexApi.init()
 
     expect(
@@ -428,7 +388,7 @@ describe('indexModels', () => {
 
   test('modelsToIndex is properly updated after indexModels()', async () => {
     const modelsToIndex = [StreamID.fromString(STREAM_ID_A), Model.MODEL]
-    const indexApi = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+    const indexApi = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
     await indexApi.init()
     expect(indexApi.getActiveModelsToIndex()).toEqual([])
     await indexApi.indexModels(modelsToIndexArgs(modelsToIndex))
@@ -445,7 +405,7 @@ describe('indexModels', () => {
 
   test('modelsToIndex is properly updated after stopIndexingModels()', async () => {
     const modelsToIndex = [StreamID.fromString(STREAM_ID_A), Model.MODEL]
-    const indexApi = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+    const indexApi = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
     await indexApi.init()
     await indexApi.indexModels(modelsToIndexArgs(modelsToIndex))
     expect(
@@ -468,42 +428,32 @@ describe('indexModels', () => {
 
 describe('indexStream', () => {
   const MODELS_TO_INDEX = [STREAM_ID_A, STREAM_ID_B].map(StreamID.fromString)
-  const STREAM_CONTENT_A = {
+  const STREAM_CONTENT = {
     model: MODELS_TO_INDEX[0],
     streamID: StreamID.fromString(STREAM_ID_B),
-    streamContent: STREAM_TEST_DATA_PROFILE_A,
-    tip: FAKE_CID_A,
     controller: CONTROLLER,
-    lastAnchor: null,
-    firstAnchor: null,
-  }
-  const STREAM_CONTENT_B = {
-    model: MODELS_TO_INDEX[0],
-    streamID: StreamID.fromString(STREAM_ID_A),
-    streamContent: STREAM_TEST_DATA_PROFILE_B,
-    tip: FAKE_CID_B,
-    controller: CONTROLLER,
+    streamContent: STREAM_TEST_DATA_PROFILE,
+    tip: FAKE_CID,
     lastAnchor: null,
     firstAnchor: null,
   }
 
-  let indexApi: PostgresIndexApi
+  let indexApi: SqliteIndexApi
   beforeEach(async () => {
-    indexApi = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+    indexApi = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
     await indexApi.init()
     await indexApi.indexModels(modelsToIndexArgs(MODELS_TO_INDEX))
   })
 
   test('new stream', async () => {
     const now = new Date()
-    await indexApi.indexStream(STREAM_CONTENT_A)
+    await indexApi.indexStream(STREAM_CONTENT)
     const result: Array<any> = await dbConnection.from(`${MODELS_TO_INDEX[0]}`).select('*')
     expect(result.length).toEqual(1)
     const raw = result[0]
     expect(raw.stream_id).toEqual(STREAM_ID_B)
     expect(raw.controller_did).toEqual(CONTROLLER)
-    expect(raw.stream_content).toEqual(STREAM_TEST_DATA_PROFILE_A)
-    expect(raw.tip).toEqual(FAKE_CID_A.toString())
+    expect(raw.stream_content).toEqual(JSON.stringify(STREAM_TEST_DATA_PROFILE))
     expect(raw.last_anchored_at).toBeNull()
     expect(raw.first_anchored_at).toBeNull()
     const createdAt = new Date(raw.created_at)
@@ -514,10 +464,10 @@ describe('indexStream', () => {
 
   test('override stream', async () => {
     const createTime = new Date()
-    await indexApi.indexStream(STREAM_CONTENT_A)
+    await indexApi.indexStream(STREAM_CONTENT)
     const updateTime = new Date(createTime.valueOf() + 5000)
     const updatedStreamContent = {
-      ...STREAM_CONTENT_A,
+      ...STREAM_CONTENT,
       updatedAt: updateTime,
       lastAnchor: updateTime,
       firstAnchor: updateTime,
@@ -538,58 +488,13 @@ describe('indexStream', () => {
     const createdAt = new Date(raw.created_at)
     expect(closeDates(createdAt, createTime)).toBeTruthy()
   })
-
-  test('verify jsonb index creation and invocation while querying', async () => {
-    await indexApi.indexStream(STREAM_CONTENT_A)
-
-    // create index on jsonb content and disable seq scans
-    await dbConnection.raw(`SET enable_seqscan = off;`)
-    let result = await dbConnection.raw(
-      `CREATE INDEX idx_postgres_jsonb ON ${STREAM_ID_A}((stream_content->'settings'->'dark_mode'))`
-    )
-    expect(result.command).toEqual('CREATE')
-
-    // query indexed jsonb data
-    result = await dbConnection.raw(
-      `EXPLAIN SELECT *
-       FROM ${MODELS_TO_INDEX[0]}
-       WHERE stream_content->'settings'->'dark_mode' = ?;`,
-      true
-    )
-    expect(result.command).toEqual('EXPLAIN')
-    expect(result.rows.length).toBeGreaterThan(1)
-    expect(result.rows[0]['QUERY PLAN'].includes('idx_postgres_jsonb')).toBeTruthy()
-  })
-
-  test('query and filter jsonb stream content', async () => {
-    await indexApi.indexStream(STREAM_CONTENT_A)
-    await indexApi.indexStream(STREAM_CONTENT_B)
-
-    let result: Array<any> = await dbConnection.select('*').from(`${MODELS_TO_INDEX[0]}`)
-    expect(result.length).toEqual(2)
-
-    // filter content from existing model per controller
-    result = await dbConnection
-      .from(`${MODELS_TO_INDEX[0]}`)
-      .select(
-        dbConnection.raw(
-          `*, stream_content->'settings'->'dark_mode' AS dark_mode, stream_content->>'id' AS id`
-        )
-      )
-      .whereRaw(`stream_content->'settings'->'dark_mode' = ?`, true)
-
-    expect(result.length).toEqual(1)
-    const raw = result[0]
-    expect(raw.dark_mode).toEqual(STREAM_TEST_DATA_PROFILE_A.settings.dark_mode)
-    expect(raw.id).toEqual(STREAM_TEST_DATA_PROFILE_A.id)
-  })
 })
 
 describe('page', () => {
   const FAUX_DB_CONNECTION = {} as unknown as Knex
 
   test('call the order if historical sync is allowed', async () => {
-    const indexApi = new PostgresIndexApi(FAUX_DB_CONNECTION, true, logger, Networks.INMEMORY)
+    const indexApi = new SqliteIndexApi(FAUX_DB_CONNECTION, true, logger, Networks.INMEMORY)
     indexApi.modelsToIndex = [StreamID.fromString(STREAM_ID_A)]
     const mockPage = jest.fn(async () => {
       return { edges: [], pageInfo: { hasNextPage: false, hasPreviousPage: false } }
@@ -599,7 +504,7 @@ describe('page', () => {
     expect(mockPage).toBeCalled()
   })
   test('throw if historical sync is not allowed', async () => {
-    const indexApi = new PostgresIndexApi(FAUX_DB_CONNECTION, false, logger, Networks.INMEMORY)
+    const indexApi = new SqliteIndexApi(FAUX_DB_CONNECTION, false, logger, Networks.INMEMORY)
     indexApi.modelsToIndex = [StreamID.fromString(STREAM_ID_A)]
     await expect(indexApi.page({ model: STREAM_ID_A, first: 100 })).rejects.toThrow(
       IndexQueryNotAvailableError
@@ -613,16 +518,14 @@ describe('count', () => {
   const MODEL = MODELS_TO_INDEX[0]
 
   test('all', async () => {
-    const indexApi = new PostgresIndexApi(dbConnection, true, logger, Networks.INMEMORY)
+    const indexApi = new SqliteIndexApi(dbConnection, true, logger, Networks.INMEMORY)
     await indexApi.init()
     await indexApi.indexModels(
       MODELS_TO_INDEX.map((m) => {
         return { model: m }
       })
     )
-    const rows = await readCsvFixture(
-      new URL('../../__tests__/insertion-order.fixture.csv', import.meta.url)
-    )
+    const rows = await readCsvFixture(new URL('./insertion-order.fixture.csv', import.meta.url))
     for (const row of rows) {
       await indexApi.indexStream(row)
     }
